requirements.txt
fastapi
uvicorn
twilio
openai
python-decouple
psycopg2-binary
python-multipart
pyngrok